// src/api/models/activity-log.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Your API
 * API description
 *
 * The version of the OpenAPI document: v1
 * Contact: dev@example.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface ActivityLog
 */
export interface ActivityLog {
    /**
     * 
     * @type {number}
     * @memberof ActivityLog
     */
    'id'?: number;
    /**
     * 
     * @type {string}
     * @memberof ActivityLog
     */
    'user'?: string;
    /**
     * 
     * @type {string}
     * @memberof ActivityLog
     */
    'action'?: ActivityLogActionEnum;
    /**
     * 
     * @type {string}
     * @memberof ActivityLog
     */
    'timestamp'?: string;
    /**
     * 
     * @type {string}
     * @memberof ActivityLog
     */
    'content_type'?: string;
    /**
     * 
     * @type {number}
     * @memberof ActivityLog
     */
    'object_id'?: number;
    /**
     * 
     * @type {string}
     * @memberof ActivityLog
     */
    'content_object'?: string;
    /**
     * 
     * @type {string}
     * @memberof ActivityLog
     */
    'project'?: string;
    /**
     * 
     * @type {string}
     * @memberof ActivityLog
     */
    'from_state'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof ActivityLog
     */
    'to_state'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof ActivityLog
     */
    'comment_text'?: string | null;
    /**
     * 
     * @type {object}
     * @memberof ActivityLog
     */
    'attachment_info'?: object | null;
}

export const ActivityLogActionEnum = {
    Create: 'create',
    Update: 'update',
    Delete: 'delete',
    Move: 'move',
    Comment: 'comment',
    Complete: 'complete',
    Add: 'add',
    StatusChange: 'status_change',
    FileUpload: 'file_upload',
    RecurrenceCreated: 'recurrence_created',
    RecurrenceTriggered: 'recurrence_triggered'
} as const;

export type ActivityLogActionEnum = typeof ActivityLogActionEnum[keyof typeof ActivityLogActionEnum];